Improved component
Changed the way the animation for opening and closing the card
is done. Now we use the Animations API instead of
requestAnimationFrame().
Improved the HTML and CSS, adding a menu tag for the buttons
inside summary that are not the toggle button of the card.
Improved the CSS for the animation of the sliding buttons in the
summary.